+++ backend/apid/routers/extensions.go
@@ -15,8 +15,8 @@ type ExtensionsRouter struct {
 	controller actions.ExtensionController
 }
 
-// NewExtensionRouter instantiates new router for controlling extension resources
-func NewExtensionRouter(store store.ExtensionRegistry) *ExtensionsRouter {
+// NewExtensionsRouter creates a new router for controlling extension resources
+func NewExtensionsRouter(store store.ExtensionRegistry) *ExtensionsRouter {
 	return &ExtensionsRouter{
 		controller: actions.NewExtensionController(store),
 	}