+++ cli/commands/commands.go
@@ -11,6 +11,7 @@ import (
 	"github.com/sensu/sensu-go/cli/commands/entity"
 	"github.com/sensu/sensu-go/cli/commands/environment"
 	"github.com/sensu/sensu-go/cli/commands/event"
+	"github.com/sensu/sensu-go/cli/commands/extension"
 	"github.com/sensu/sensu-go/cli/commands/filter"
 	"github.com/sensu/sensu-go/cli/commands/handler"
 	"github.com/sensu/sensu-go/cli/commands/hook"
@@ -48,6 +49,7 @@ func AddCommands(rootCmd *cobra.Command, cli *cli.SensuCli) {
 		user.HelpCommand(cli),
 		silenced.HelpCommand(cli),
 		create.CreateCommand(cli),
+		extension.HelpCommand(cli),
 	)
 
 	for _, cmd := range rootCmd.Commands() {